feat: 주문내역 레이아웃 분리 (#56)
